pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic-claw-router"
version = "0.1.0"
description = "System-level intelligent router for LLM Mixture-of-Models"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "ruff>=0.5.0",
]

[project.scripts]
semantic-claw-router = "semantic_claw_router.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests requiring live endpoints",
    "benchmark: marks benchmark tests",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.setuptools.packages.find]
where = ["src"]
